# frontend/generated/openapi-current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/behaviour/profile": {
      "get": {
        "tags": [
          "behaviour"
        ],
        "summary": "Get Financial Profile",
        "description": "Get comprehensive financial behaviour profile.\n\nReturns financial personality classification based on transaction and account data.\n\nArgs:\n    household_id: Household identifier (default: \"default\")\n\nReturns:\n    FinancialProfileResponse with profile classification",
        "operationId": "get_financial_profile_api_v1_behaviour_profile_get",
        "parameters": [
          {
            "name": "household_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Household identifier",
              "default": "default",
              "title": "Household Id"
            },
            "description": "Household identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/behaviour/wellness-score": {
      "get": {
        "tags": [
          "behaviour"
        ],
        "summary": "Get Wellness Score",
        "description": "Get the latest financial wellness score.\n\nReturns wellness score with band classification and component breakdown.\n\nArgs:\n    household_id: Household identifier (default: \"default\")\n\nReturns:\n    WellnessScoreResponse with score, band, and components",
        "operationId": "get_wellness_score_api_v1_behaviour_wellness_score_get",
        "parameters": [
          {
            "name": "household_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Household identifier",
              "default": "default",
              "title": "Household Id"
            },
            "description": "Household identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WellnessScoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/behaviour/patterns": {
      "get": {
        "tags": [
          "behaviour"
        ],
        "summary": "Get Patterns",
        "description": "Get detected financial patterns.\n\nReturns patterns like impulse spending and subscriptions with strength scores.\n\nArgs:\n    household_id: Household identifier (default: \"default\")\n    pattern_type: Optional filter for specific pattern type\n    days: Number of days to look back (1-365, default: 30)\n\nReturns:\n    List of FinancialPattern objects",
        "operationId": "get_patterns_api_v1_behaviour_patterns_get",
        "parameters": [
          {
            "name": "household_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Household identifier",
              "default": "default",
              "title": "Household Id"
            },
            "description": "Household identifier"
          },
          {
            "name": "pattern_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by pattern type (e.g., IMPULSE, SUBSCRIPTION)",
              "title": "Pattern Type"
            },
            "description": "Filter by pattern type (e.g., IMPULSE, SUBSCRIPTION)"
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "Number of days to look back for patterns",
              "default": 30,
              "title": "Days"
            },
            "description": "Number of days to look back for patterns"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Patterns Api V1 Behaviour Patterns Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/behaviour/recommendations": {
      "get": {
        "tags": [
          "behaviour"
        ],
        "summary": "Get Recommendations",
        "description": "Get financial recommendations based on behaviour metrics.\n\nReturns actionable recommendations sorted by severity.\n\nArgs:\n    household_id: Household identifier (default: \"default\")\n    limit: Maximum number of recommendations to return (1-50, default: 10)\n    severity: Optional filter for specific severity level\n\nReturns:\n    RecommendationsResponse with triggered recommendations",
        "operationId": "get_recommendations_api_v1_behaviour_recommendations_get",
        "parameters": [
          {
            "name": "household_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Household identifier",
              "default": "default",
              "title": "Household Id"
            },
            "description": "Household identifier"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Maximum number of recommendations to return",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum number of recommendations to return"
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by severity (LOW, MEDIUM, HIGH, CRITICAL)",
              "title": "Severity"
            },
            "description": "Filter by severity (LOW, MEDIUM, HIGH, CRITICAL)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/behaviour/monthly-report": {
      "get": {
        "tags": [
          "behaviour"
        ],
        "summary": "Get Monthly Report",
        "description": "Generate a monthly financial summary report.\n\nReturns comprehensive summary including wellness, debt, cashflow, patterns, and alerts.\n\nArgs:\n    period: Period in YYYY-MM format (default: current month)\n    household_id: Household identifier (default: \"default\")\n\nReturns:\n    MonthlySummaryResponse with comprehensive financial summary",
        "operationId": "get_monthly_report_api_v1_behaviour_monthly_report_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Period in YYYY-MM format (default: current month)",
              "title": "Period"
            },
            "description": "Period in YYYY-MM format (default: current month)"
          },
          {
            "name": "household_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Household identifier",
              "default": "default",
              "title": "Household Id"
            },
            "description": "Household identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlySummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/behaviour": {
      "get": {
        "tags": [
          "behaviour-workspace"
        ],
        "summary": "Get Behaviour",
        "description": "Get behaviour summary for the Behaviour Intelligence Workspace.\n\nReturns aggregated data matching BehaviourViewModel format.",
        "operationId": "get_behaviour_api_v1_behaviour_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "monthly",
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Behaviour Api V1 Behaviour Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CashflowHealthResponse": {
        "properties": {
          "cashflow_stability_index": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cashflow Stability Index",
            "description": "Cashflow stability index (0-1)"
          },
          "income_stability": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Income Stability",
            "description": "Income stability score (0-1)"
          },
          "expense_stability": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Expense Stability",
            "description": "Expense stability score (0-1)"
          },
          "monthly_surplus_paise": {
            "type": "integer",
            "title": "Monthly Surplus Paise",
            "description": "Monthly surplus in paise (can be negative)"
          },
          "snapshot_date": {
            "type": "string",
            "title": "Snapshot Date",
            "description": "Date of the snapshot in ISO format"
          }
        },
        "type": "object",
        "required": [
          "cashflow_stability_index",
          "income_stability",
          "expense_stability",
          "monthly_surplus_paise",
          "snapshot_date"
        ],
        "title": "CashflowHealthResponse",
        "description": "Response model for cashflow health metrics."
      },
...
      "DebtHealthResponse": {
        "properties": {
          "foir": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Foir",
            "description": "Fixed Obligation to Income Ratio (0-1)"
          },
          "credit_dependency_ratio": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Credit Dependency Ratio",
            "description": "Ratio of credit-funded expenses to total expenses (0-1+)"
          },
          "debt_cycle_score": {
            "type": "integer",
            "title": "Debt Cycle Score",
            "description": "Debt cycle score (0-100, higher = worse)"
          },
          "credit_revolver_ratio": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Credit Revolver Ratio",
            "description": "Ratio of revolving credit usage (0-1)"
          },
          "band": {
            "type": "string",
            "enum": [
              "HEALTHY",
              "MODERATE",
              "WARNING",
              "DANGER"
            ],
            "title": "Band",
            "description": "Debt health classification band"
          },
          "snapshot_date": {
            "type": "string",
            "title": "Snapshot Date",
            "description": "Date of the snapshot in ISO format"
          }
        },
        "type": "object",
        "required": [
          "foir",
          "credit_dependency_ratio",
          "debt_cycle_score",
          "credit_revolver_ratio",
          "band",
          "snapshot_date"
        ],
        "title": "DebtHealthResponse",
        "description": "Response model for debt health metrics."
      },
...
      "FinancialProfileResponse": {
        "properties": {
          "profile_type": {
            "type": "string",
            "enum": [
              "SAVER",
              "BALANCED",
              "SPENDER",
              "DEBT_OPTIMIZER",
              "DEBT_DEPENDENT",
              "INSUFFICIENT_DATA"
            ],
            "title": "Profile Type",
            "description": "Financial personality profile"
          },
          "confidence": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence",
            "description": "Confidence score (0-1)"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "description": "Explanation of the profile classification"
          },
          "snapshot_date": {
            "type": "string",
            "title": "Snapshot Date",
            "description": "Date of the snapshot in ISO format"
          }
        },
        "type": "object",
        "required": [
          "profile_type",
          "confidence",
          "explanation",
          "snapshot_date"
        ],
        "title": "FinancialProfileResponse",
        "description": "Response model for financial personality profile."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MonthlySummaryResponse": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period",
            "description": "Summary period (YYYY-MM)"
          },
          "wellness_score": {
            "$ref": "#/components/schemas/WellnessScoreResponse",
            "description": "Wellness score for the period"
          },
          "debt_health": {
            "$ref": "#/components/schemas/DebtHealthResponse",
            "description": "Debt health metrics"
          },
          "cashflow_health": {
            "$ref": "#/components/schemas/CashflowHealthResponse",
            "description": "Cashflow health metrics"
          },
          "top_patterns": {
            "items": {
              "$ref": "#/components/schemas/FinancialPattern"
            },
            "type": "array",
            "title": "Top Patterns",
            "description": "Top detected financial patterns"
          },
          "savings_rate": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Savings Rate",
            "description": "Savings rate for the period"
          },
          "total_income_paise": {
            "type": "integer",
            "title": "Total Income Paise",
            "description": "Total income in paise"
          },
          "total_expenses_paise": {
            "type": "integer",
            "title": "Total Expenses Paise",
            "description": "Total expenses in paise"
          },
          "alerts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alerts",
            "description": "Financial alerts for the period"
          }
        },
        "type": "object",
        "required": [
          "period",
          "wellness_score",
          "debt_health",
          "cashflow_health",
          "top_patterns",
          "savings_rate",
          "total_income_paise",
          "total_expenses_paise",
          "alerts"
        ],
        "title": "MonthlySummaryResponse",
        "description": "Response model for monthly financial summary."
      },
...
      "RecommendationResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short title of the recommendation"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Human-readable explanation of the recommendation"
          },
          "metric": {
            "type": "string",
            "title": "Metric",
            "description": "The metric value that triggered this recommendation"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "description": "Severity level (LOW, MEDIUM, HIGH, CRITICAL)"
          },
          "suggested_action": {
            "type": "string",
            "title": "Suggested Action",
            "description": "Actionable suggestion for the user"
          }
        },
        "type": "object",
        "required": [
          "title",
          "reason",
          "metric",
          "severity",
          "suggested_action"
        ],
        "title": "RecommendationResponse",
        "description": "Response model for a single recommendation."
      },
      "RecommendationsResponse": {
        "properties": {
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendationResponse"
            },
            "type": "array",
            "title": "Recommendations",
            "description": "List of triggered recommendations sorted by severity"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Total number of recommendations"
          },
          "snapshot_date": {
            "type": "string",
            "title": "Snapshot Date",
            "description": "Date of the snapshot used for recommendations"
          }
        },
        "type": "object",
        "required": [
          "recommendations",
          "total_count",
          "snapshot_date"
        ],
        "title": "RecommendationsResponse",
        "description": "Response model for a list of recommendations."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WellnessScoreResponse": {
        "properties": {
          "score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Score",
            "description": "Wellness score between 0 and 100"
          },
          "band": {
            "type": "string",
            "enum": [
              "Excellent",
              "Healthy",
              "Developing",
              "Risk",
              "Critical"
            ],
            "title": "Band",
            "description": "Wellness classification band"
          },
          "components": {
            "additionalProperties": {
              "type": "string",
              "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
            },
            "type": "object",
            "title": "Components",
            "description": "Breakdown of wellness score components"
          },
          "snapshot_date": {
            "type": "string",
            "title": "Snapshot Date",
            "description": "Date of the snapshot in ISO format"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "description": "Version of the scoring algorithm"
          }
        },
        "type": "object",
        "required": [
          "score",
          "band",
          "components",
          "snapshot_date",
          "version"
        ],
        "title": "WellnessScoreResponse",
        "description": "Response model for wellness score."
      }
    }
  }
}